requirements.txt
kafka==1.3.5
PyMySQL==0.9.3
python-dateutil==2.8.1